+++ source/common.c
@@ -207,7 +207,7 @@ int clean_directory(const char* inputdir)
 	{
 		if (strcmp(dir->d_name, ".") != 0 && strcmp(dir->d_name, "..") != 0)
 		{
-			snprintf(dataPath, sizeof(dataPath), "%s" "%s", inputdir, dir->d_name);
+			snprintf(dataPath, sizeof(dataPath), "%s%s", inputdir, dir->d_name);
 			unlink_secure(dataPath);
 		}
 	}