+++ source/menu_cheats.c
@@ -309,7 +309,7 @@ void DrawGameList(int selIndex, list_t * games, u8 alpha)
     
     list_node_t *node;
     game_entry_t *item;
-//    char tmp[4] = "   ";
+    char tmp[4] = "   ";
     int game_y = help_png_y, y_inc = APP_LINE_OFFSET;
     int maxPerPage = (SCREEN_HEIGHT - (game_y * 2) - 30) / y_inc;
     
@@ -346,15 +346,15 @@ void DrawGameList(int selIndex, list_t * games, u8 alpha)
 			if (item->title_id)
 				DrawString((SCREEN_WIDTH - 40) - (MENU_ICON_OFF * 5), game_y, item->title_id);
 
-/*
 			tmp[0] = ' ';
-			if (item->flags & CHEAT_FLAG_PS1) tmp[0] = CHAR_TAG_PS1;
-			if (item->flags & CHEAT_FLAG_PS2) tmp[0] = CHAR_TAG_PS2;
-			if (item->flags & CHEAT_FLAG_PSP) tmp[0] = CHAR_TAG_PSP;
-			if (item->flags & CHEAT_FLAG_PS4) tmp[0] = CHAR_TAG_PS4;
+			if (item->flags & CHEAT_FLAG_SHN) tmp[0] = CHAR_TAG_SHN;
+			if (item->flags & CHEAT_FLAG_MC4) tmp[0] = CHAR_TAG_MC4;
+			if (item->flags & CHEAT_FLAG_JSON) tmp[0] = CHAR_TAG_JSON;
+/*
 			tmp[1] = (item->flags & CHEAT_FLAG_OWNER) ? CHAR_TAG_OWNER : ' ';
 			tmp[2] = (item->flags & CHEAT_FLAG_LOCKED) ? CHAR_TAG_LOCKED : ' ';
 */
+			DrawString(SCREEN_WIDTH - (MENU_ICON_OFF * 2), game_y, tmp);
 
 			if (item->flags & CHEAT_FLAG_OWNER)
 				DrawString(MENU_ICON_OFF + MENU_TITLE_OFF - 50, game_y, "\xE2\x98\x85");
@@ -398,19 +398,16 @@ void DrawCheatsList(int selIndex, game_entry_t* game, u8 alpha)
 
     for (; x < max; x++)
     {
-        int xo = 0; //(((selIndex - x) < 0) ? -(selIndex - x) : (selIndex - x));
-        
         if (x >= 0 && node)
         {
 			code = list_get(node);
-            //u32 color = game.codes[x].activated ? 0x4040C000 : 0x00000000;
-			u8 a = (u8)((alpha * CalculateAlphaList(x, selIndex, maxPerPage)) / 0xFF);
+			u8 a = (u8)((alpha * CalculateAlphaList(x, selIndex, maxPerPage - 1)) / 0xFF);
 
             if (!a)
                 goto skip_code;
 
             SetFontColor(APP_FONT_COLOR | a, 0);
-            float dx = DrawString(MENU_ICON_OFF + (MENU_TITLE_OFF * 3) - xo, game_y, code->name);
+            float dx = DrawString(MENU_ICON_OFF + (MENU_TITLE_OFF * 3), game_y, code->name);
             
             if (code->activated)
             {